# internal/app/adapter/server/http/api/docs/swagger.yaml
basePath: /api
consumes:
- application/json
- application/x-www-form-urlencoded
definitions:
  example.ClientError:
    properties:
      errMsg:
        example: 客户端请求错误|参数校验错误
        type: string
      errNo:
        description: errNo 类型应为 int，string 仅为了表达多个错误码
        example: 10002|10003
        type: string
    type: object
  example.PermissionDenied:
    properties:
      errMsg:
        example: 暂无权限
        type: string
      errNo:
        example: 10005
        type: integer
    type: object
  example.ResourceNotFound:
    properties:
      errMsg:
        example: 资源不存在
        type: string
      errNo:
        example: 10006
        type: integer
    type: object
  example.ServerError:
    properties:
      errMsg:
        example: 服务器出错
        type: string
      errNo:
        example: 10001
        type: integer
    type: object
  example.Success:
    properties:
      data: {}
    type: object
  example.TooManyRequest:
    properties:
      errMsg:
        example: 请求过于频繁
        type: string
      errNo:
        example: 10007
        type: integer
    type: object
  example.Unauthorized:
    properties:
      errMsg:
        example: 未经授权
        type: string
      errNo:
        example: 10004
        type: integer
    type: object
...
  v1.SystemRoleCreateRequest:
    properties:
      name:
        type: string
    type: object
  v1.SystemRoleDetailResponse:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  v1.SystemRoleGetPermissionsRequest:
    properties:
      id:
        type: integer
    type: object
  v1.SystemRoleGrantPermissionsRequest:
    properties:
      permissions:
        items:
          type: integer
        type: array
      role:
        type: integer
    type: object
  v1.SystemRoleInfo:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  v1.SystemRoleUpdateRequest:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
host: localhost
info:
  contact: {}
  description: API 接口文档
  title: API 接口文档
  version: 0.0.0
paths:
...
  /v1/system-role:
    post:
      consumes:
      - application/json
      description: 角色创建
      parameters:
      - description: 权限信息
        format: string
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/v1.SystemRoleCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/example.Success'
        "400":
          description: 客户端请求错误（code 类型应为 int，string 仅为了表达多个错误码）
          schema:
            $ref: '#/definitions/example.ClientError'
        "401":
          description: 登陆失效
          schema:
            $ref: '#/definitions/example.Unauthorized'
        "403":
          description: 没有权限
          schema:
            $ref: '#/definitions/example.PermissionDenied'
        "404":
          description: 资源不存在
          schema:
            $ref: '#/definitions/example.ResourceNotFound'
        "429":
          description: 请求过于频繁
          schema:
            $ref: '#/definitions/example.TooManyRequest'
        "500":
          description: 服务器出错
          schema:
            $ref: '#/definitions/example.ServerError'
      security:
      - Authorization: []
      summary: 角色创建
      tags:
      - 角色
    put:
      consumes:
      - application/json
      description: 角色更新
      parameters:
      - description: 权限信息
        format: string
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/v1.SystemRoleUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/example.Success'
        "400":
          description: 客户端请求错误（code 类型应为 int，string 仅为了表达多个错误码）
          schema:
            $ref: '#/definitions/example.ClientError'
        "401":
          description: 登陆失效
          schema:
            $ref: '#/definitions/example.Unauthorized'
        "403":
          description: 没有权限
          schema:
            $ref: '#/definitions/example.PermissionDenied'
        "404":
          description: 资源不存在
          schema:
            $ref: '#/definitions/example.ResourceNotFound'
        "429":
          description: 请求过于频繁
          schema:
            $ref: '#/definitions/example.TooManyRequest'
        "500":
          description: 服务器出错
          schema:
            $ref: '#/definitions/example.ServerError'
      security:
      - Authorization: []
      summary: 角色更新
      tags:
      - 角色
  /v1/system-role/{id}:
    delete:
      consumes:
      - text/plain
      description: 角色删除
      parameters:
      - description: 角色 id
        format: uint
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/example.Success'
        "400":
          description: 客户端请求错误（code 类型应为 int，string 仅为了表达多个错误码）
          schema:
            $ref: '#/definitions/example.ClientError'
        "401":
          description: 登陆失效
          schema:
            $ref: '#/definitions/example.Unauthorized'
        "403":
          description: 没有权限
          schema:
            $ref: '#/definitions/example.PermissionDenied'
        "404":
          description: 资源不存在
          schema:
            $ref: '#/definitions/example.ResourceNotFound'
        "429":
          description: 请求过于频繁
          schema:
            $ref: '#/definitions/example.TooManyRequest'
        "500":
          description: 服务器出错
          schema:
            $ref: '#/definitions/example.ServerError'
      security:
      - Authorization: []
      summary: 角色删除
      tags:
      - 角色
    get:
      consumes:
      - text/plain
      description: 角色详情
      parameters:
      - description: 角色 id
        format: uint
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            allOf:
            - $ref: '#/definitions/example.Success'
            - properties:
                data:
                  $ref: '#/definitions/v1.SystemRoleDetailResponse'
              type: object
        "400":
          description: 客户端请求错误（code 类型应为 int，string 仅为了表达多个错误码）
          schema:
            $ref: '#/definitions/example.ClientError'
        "401":
          description: 登陆失效
          schema:
            $ref: '#/definitions/example.Unauthorized'
        "403":
          description: 没有权限
          schema:
            $ref: '#/definitions/example.PermissionDenied'
        "404":
          description: 资源不存在
          schema:
            $ref: '#/definitions/example.ResourceNotFound'
        "429":
          description: 请求过于频繁
          schema:
            $ref: '#/definitions/example.TooManyRequest'
        "500":
          description: 服务器出错
          schema:
            $ref: '#/definitions/example.ServerError'
      security:
      - Authorization: []
      summary: 角色详情
      tags:
      - 角色
  /v1/system-role/permissions:
    get:
      consumes:
      - application/json
      description: 获取角色权限
      parameters:
      - description: 请求体
        format: string
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/v1.SystemRoleGetPermissionsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/example.Success'
        "400":
          description: 客户端请求错误（code 类型应为 int，string 仅为了表达多个错误码）
          schema:
            $ref: '#/definitions/example.ClientError'
        "401":
          description: 登陆失效
          schema:
            $ref: '#/definitions/example.Unauthorized'
        "403":
          description: 没有权限
          schema:
            $ref: '#/definitions/example.PermissionDenied'
        "404":
          description: 资源不存在
          schema:
            $ref: '#/definitions/example.ResourceNotFound'
        "429":
          description: 请求过于频繁
          schema:
            $ref: '#/definitions/example.TooManyRequest'
        "500":
          description: 服务器出错
          schema:
            $ref: '#/definitions/example.ServerError'
      security:
      - Authorization: []
      summary: 获取角色权限
      tags:
      - 角色
    post:
      consumes:
      - application/json
      description: 授权角色权限
      parameters:
      - description: 请求体
        format: string
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/v1.SystemRoleGrantPermissionsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/example.Success'
        "400":
          description: 客户端请求错误（code 类型应为 int，string 仅为了表达多个错误码）
          schema:
            $ref: '#/definitions/example.ClientError'
        "401":
          description: 登陆失效
          schema:
            $ref: '#/definitions/example.Unauthorized'
        "403":
          description: 没有权限
          schema:
            $ref: '#/definitions/example.PermissionDenied'
        "404":
          description: 资源不存在
          schema:
            $ref: '#/definitions/example.ResourceNotFound'
        "429":
          description: 请求过于频繁
          schema:
            $ref: '#/definitions/example.TooManyRequest'
        "500":
          description: 服务器出错
          schema:
            $ref: '#/definitions/example.ServerError'
      security:
      - Authorization: []
      summary: 授权角色权限
      tags:
      - 角色
  /v1/system-roles:
    get:
      consumes:
      - application/x-www-form-urlencoded
      description: 角色列表
      parameters:
      - description: 查询字符串
        format: string
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            allOf:
            - $ref: '#/definitions/example.Success'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/v1.SystemRoleInfo'
                  type: array
              type: object
        "400":
          description: 客户端请求错误（code 类型应为 int，string 仅为了表达多个错误码）
          schema:
            $ref: '#/definitions/example.ClientError'
        "401":
          description: 登陆失效
          schema:
            $ref: '#/definitions/example.Unauthorized'
        "403":
          description: 没有权限
          schema:
            $ref: '#/definitions/example.PermissionDenied'
        "404":
          description: 资源不存在
          schema:
            $ref: '#/definitions/example.ResourceNotFound'
        "429":
          description: 请求过于频繁
          schema:
            $ref: '#/definitions/example.TooManyRequest'
        "500":
          description: 服务器出错
          schema:
            $ref: '#/definitions/example.ServerError'
      security:
      - Authorization: []
      summary: 列表
      tags:
      - 角色
...
produces:
- application/json
schemes:
- http
- https
securityDefinitions:
  Authorization:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"